fix(inflection): harden Unicode generation and structure
