+++ src/Humanizer/Inflections/InflectionEngine.cs
@@ -1,442 +1,6 @@
 namespace Humanizer;
 
-enum InflectionStatus
-{
-    Exact,
-    Productive,
-    Invariant,
-    Ambiguous,
-    Unsupported,
-    Unknown
-}
-
-enum InflectionDirection
-{
-    Forward,
-    Reverse
-}
-
-enum InflectionCasing
-{
-    Exact,
-    LowerTitleUpper,
-    None
-}
-
-enum InflectionCapability
-{
-    DisplayByCategory,
-    Invariant
-}
-
-enum InflectionQuantitySelector
-{
-    None,
-    ExactNumericSingleton
-}
-
-[Flags]
-enum InflectionCountability : byte
-{
-    None = 0,
-    Count = 1 << 0,
-    Mass = 1 << 1,
-    Collective = 1 << 2,
-    PluralOnly = 1 << 3,
-    All = Count | Mass | Collective | PluralOnly
-}
-
-readonly struct InflectionQuantity
-{
-    readonly CardinalPluralOperands operands;
-    readonly decimal decimalValue;
-    readonly double doubleValue;
-    readonly InflectionQuantityKind kind;
-
-    public InflectionQuantity(CardinalPluralOperands operands)
-    {
-        this.operands = operands;
-        decimalValue = default;
-        doubleValue = default;
-        kind = InflectionQuantityKind.Operands;
-    }
-
-    public InflectionQuantity(decimal value)
-    {
-        operands = default;
-        decimalValue = value;
-        doubleValue = default;
-        kind = InflectionQuantityKind.Decimal;
-    }
-
-    public InflectionQuantity(double value)
-    {
-        operands = default;
-        decimalValue = default;
-        doubleValue = value;
-        kind = InflectionQuantityKind.Double;
-    }
-
-    public bool HasValue => kind != InflectionQuantityKind.None;
-
-    public bool IsFinite =>
-        kind switch
-        {
-            InflectionQuantityKind.Operands => operands.IsSupported,
-            InflectionQuantityKind.Double => double.IsFinite(doubleValue),
-            _ => true
-        };
-
-    public bool IsExactNumericSingleton =>
-        kind switch
-        {
-            InflectionQuantityKind.Operands => operands.IntegerDigits.IsOne && operands.FractionDigits.IsZero,
-            InflectionQuantityKind.Decimal => decimalValue is 1m or -1m,
-            InflectionQuantityKind.Double => doubleValue is 1d or -1d,
-            _ => false
-        };
-}
-
-enum InflectionQuantityKind : byte
-{
-    None,
-    Operands,
-    Decimal,
-    Double
-}
-
-readonly struct InflectionResult(InflectionStatus status, string value)
-{
-    public InflectionStatus Status { get; } = status;
-    public string Value { get; } = value;
-}
-
-readonly struct InflectionLexemeDisplay(
-    CardinalPluralCategory category,
-    int preferredEntryIndex)
-{
-    public CardinalPluralCategory Category { get; } = category;
-    public int PreferredEntryIndex { get; } = preferredEntryIndex;
-}
-
-readonly struct InflectionLexemeRecord(
-    string id,
-    int singularEntryIndex,
-    int dictionaryPluralEntryIndex,
-    InflectionLexemeDisplay[] display,
-    InflectionCountability countability)
-{
-    public string Id { get; } = id;
-    public int SingularEntryIndex { get; } = singularEntryIndex;
-    public int DictionaryPluralEntryIndex { get; } = dictionaryPluralEntryIndex;
-    public InflectionLexemeDisplay[] Display { get; } = display;
-    public InflectionCountability Countability { get; } = countability;
-
-    public bool TryGetDisplay(
-        CardinalPluralCategory category,
-        out int entryIndex)
-    {
-        foreach (var form in Display)
-        {
-            if (form.Category == category)
-            {
-                entryIndex = form.PreferredEntryIndex;
-                return true;
-            }
-        }
-
-        entryIndex = -1;
-        return false;
-    }
-}
-
-readonly struct InflectionLexemeEntry(
-    string value,
-    int candidateOffset,
-    int candidateCount)
-{
-    public string Value { get; } = value;
-    public int CandidateOffset { get; } = candidateOffset;
-    public int CandidateCount { get; } = candidateCount;
-}
-
-readonly struct InflectionLexemeCandidate(
-    int lexemeIndex,
-    InflectionExactRole roles)
-{
-    public int LexemeIndex { get; } = lexemeIndex;
-    public InflectionExactRole Roles { get; } = roles;
-}
-
-readonly struct InflectionEntryIndexes
-{
-    readonly ushort[]? compact;
-    readonly int[]? wide;
-
-    public InflectionEntryIndexes(ushort[] compact)
-    {
-        this.compact = compact;
-        wide = null;
-    }
-
-    public InflectionEntryIndexes(int[] wide)
-    {
-        compact = null;
-        this.wide = wide;
-    }
-
-    public int Length => compact?.Length ?? wide!.Length;
-
-    public int this[int index] =>
-        compact is null
-            ? wide![index]
-            : compact[index];
-
-    public bool Contains(int value)
-    {
-        for (var index = 0; index < Length; index++)
-        {
-            if (this[index] == value)
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-}
-
-readonly struct InflectionRuleDisplay(
-    CardinalPluralCategory category,
-    string template)
-{
-    public CardinalPluralCategory Category { get; } = category;
-    public string Template { get; } = template;
-}
-
-
-readonly struct InflectionRule
-{
-    public InflectionRule(
-        string id,
-        InflectionDirection direction,
-        int priority,
-        string suffix,
-        string dictionaryPlural,
-        bool reverseEnabled,
-        bool requiresExistingLexeme,
-        InflectionUnicodeScripts scripts = InflectionUnicodeScripts.All,
-        InflectionCountability countabilities = InflectionCountability.All)
-        : this(
-            id,
-            direction,
-            priority,
-            prefix: string.Empty,
-            suffix,
-            precedingNot: [],
-            dictionaryPlural,
-            display: [],
-            excludedSurfaces: [],
-            excludedLexemes: Array.Empty<ushort>(),
-            reverseEnabled,
-            requiresExistingLexeme,
-            scripts,
-            countabilities)
-    {
-    }
-
-    public InflectionRule(
-        string id,
-        InflectionDirection direction,
-        int priority,
-        string prefix,
-        string suffix,
-        string[] precedingNot,
-        string dictionaryPlural,
-        InflectionRuleDisplay[] display,
-        string[] excludedSurfaces,
-        bool reverseEnabled,
-        bool requiresExistingLexeme,
-        InflectionUnicodeScripts scripts = InflectionUnicodeScripts.All,
-        InflectionCountability countabilities = InflectionCountability.All)
-        : this(
-            id,
-            direction,
-            priority,
-            prefix,
-            suffix,
-            precedingNot,
-            dictionaryPlural,
-            display,
-            excludedSurfaces,
-            excludedLexemes: Array.Empty<ushort>(),
-            reverseEnabled,
-            requiresExistingLexeme,
-            scripts,
-            countabilities)
-    {
-    }
-
-    public InflectionRule(
-        string id,
-        InflectionDirection direction,
-        int priority,
-        string prefix,
-        string suffix,
-        string[] precedingNot,
-        string dictionaryPlural,
-        InflectionRuleDisplay[] display,
-        string[] excludedSurfaces,
-        ushort[] excludedLexemes,
-        bool reverseEnabled,
-        bool requiresExistingLexeme,
-        InflectionUnicodeScripts scripts = InflectionUnicodeScripts.All,
-        InflectionCountability countabilities = InflectionCountability.All)
-        : this(
-            id,
-            direction,
-            priority,
-            prefix,
-            suffix,
-            precedingNot,
-            dictionaryPlural,
-            display,
-            excludedSurfaces,
-            new InflectionEntryIndexes(excludedLexemes),
-            reverseEnabled,
-            requiresExistingLexeme,
-            scripts,
-            countabilities)
-    {
-    }
-
-    public InflectionRule(
-        string id,
-        InflectionDirection direction,
-        int priority,
-        string prefix,
-        string suffix,
-        string[] precedingNot,
-        string dictionaryPlural,
-        InflectionRuleDisplay[] display,
-        string[] excludedSurfaces,
-        int[] excludedLexemes,
-        bool reverseEnabled,
-        bool requiresExistingLexeme,
-        InflectionUnicodeScripts scripts = InflectionUnicodeScripts.All,
-        InflectionCountability countabilities = InflectionCountability.All)
-        : this(
-            id,
-            direction,
-            priority,
-            prefix,
-            suffix,
-            precedingNot,
-            dictionaryPlural,
-            display,
-            excludedSurfaces,
-            new InflectionEntryIndexes(excludedLexemes),
-            reverseEnabled,
-            requiresExistingLexeme,
-            scripts,
-            countabilities)
-    {
-    }
-
-    InflectionRule(
-        string id,
-        InflectionDirection direction,
-        int priority,
-        string prefix,
-        string suffix,
-        string[] precedingNot,
-        string dictionaryPlural,
-        InflectionRuleDisplay[] display,
-        string[] excludedSurfaces,
-        InflectionEntryIndexes excludedLexemes,
-        bool reverseEnabled,
-        bool requiresExistingLexeme,
-        InflectionUnicodeScripts scripts,
-        InflectionCountability countabilities)
-    {
-        Id = id;
-        Direction = direction;
-        Priority = priority;
-        Prefix = prefix;
-        Suffix = suffix;
-        PrecedingNot = precedingNot;
-        DictionaryPlural = dictionaryPlural;
-        Display = display;
-        ExcludedSurfaces = excludedSurfaces;
-        ExcludedLexemes = excludedLexemes;
-        ReverseEnabled = reverseEnabled;
-        RequiresExistingLexeme = requiresExistingLexeme;
-        Scripts = scripts;
-        Countabilities = countabilities;
-    }
-
-    public string Id { get; }
-    public InflectionDirection Direction { get; }
-    public int Priority { get; }
-    public string Prefix { get; }
-    public string Suffix { get; }
-    public string[] PrecedingNot { get; }
-    public string DictionaryPlural { get; }
-    public InflectionRuleDisplay[] Display { get; }
-    public string[] ExcludedSurfaces { get; }
-    public InflectionEntryIndexes ExcludedLexemes { get; }
-    public bool ReverseEnabled { get; }
-    public bool RequiresExistingLexeme { get; }
-    public InflectionUnicodeScripts Scripts { get; }
-    public InflectionCountability Countabilities { get; }
-
-    public bool SupportsScripts(InflectionUnicodeScripts detectedScripts) =>
-        (Scripts & detectedScripts) != InflectionUnicodeScripts.None;
-
-    public bool SupportsCountability(InflectionCountability countability) =>
-        countability == InflectionCountability.None
-            ? Countabilities == InflectionCountability.All
-            : (Countabilities & countability) != InflectionCountability.None;
-
-    public bool TryGetTemplate(
-        CardinalPluralCategory? category,
-        [NotNullWhen(true)] out string? template)
-    {
-        if (category is null)
-        {
-            template = DictionaryPlural;
-            return true;
-        }
-
-        foreach (var form in Display)
-        {
-            if (form.Category == category)
-            {
-                template = form.Template;
-                return true;
-            }
-        }
-
-        template = null;
-        return false;
-    }
-}
-
-[Flags]
-enum InflectionExactRole : byte
-{
-    None = 0,
-    Singular = 1,
-    DictionaryPlural = 2,
-    Zero = 4,
-    One = 8,
-    Two = 16,
-    Few = 32,
-    Many = 64,
-    Other = 128
-}
-
-sealed class InflectionBundle
+sealed partial class InflectionBundle
 {
     readonly InflectionCasing casing;
     readonly InflectionQuantitySelector quantitySelector;
@@ -1012,274 +576,6 @@ sealed class InflectionBundle
         return new(InflectionStatus.Unknown, originalInput);
     }
 
-    InflectionResult Project(
-        string input,
-        string output,
-        CaseProjection projection,
-        InflectionStatus status)
-    {
-        if (output.Length == 0 || !IsWellFormedUtf16(output))
-        {
-            return new(InflectionStatus.Unsupported, input);
-        }
-
-        string projected;
-        if (casing == InflectionCasing.LowerTitleUpper &&
-            projection == CaseProjection.Title)
-        {
-            if (!TryProjectUpper(output, firstOnly: true, out projected))
-            {
-                return new(InflectionStatus.Unsupported, input);
-            }
-        }
-        else if (casing == InflectionCasing.LowerTitleUpper &&
-                 projection == CaseProjection.Upper)
-        {
-            if (!TryProjectUpper(output, firstOnly: false, out projected))
-            {
-                return new(InflectionStatus.Unsupported, input);
-            }
-        }
-        else
-        {
-            projected = output;
-        }
-
-        if (!IsScriptEligible(projected, allowNonLetters: true))
-        {
-            return new(InflectionStatus.Unsupported, input);
-        }
-
-        return string.Equals(projected, input, StringComparison.Ordinal)
-            ? new(InflectionStatus.Invariant, input)
-            : new(status, projected);
-    }
-
-    static bool TryProjectUpper(
-        string value,
-        bool firstOnly,
-        out string projected)
-    {
-        char[]? characters = null;
-        for (var index = 0; index < value.Length;)
-        {
-            var scalarOffset = index;
-            var scalar = ReadScalar(value, ref index, value.Length);
-            var upper = InflectionUnicodeData.ToUpperSimple(scalar);
-            if (upper != scalar)
-            {
-                var scalarLength = index - scalarOffset;
-                var upperLength = upper <= char.MaxValue ? 1 : 2;
-                if (upperLength != scalarLength)
-                {
-                    projected = value;
-                    return false;
-                }
-
-                characters ??= value.ToCharArray();
-                if (upperLength == 1)
-                {
-                    characters[scalarOffset] = (char)upper;
-                }
-                else
-                {
-                    var supplementary = upper - 0x10000;
-                    characters[scalarOffset] =
-                        (char)((supplementary / 0x400) + 0xD800);
-                    characters[scalarOffset + 1] =
-                        (char)((supplementary % 0x400) + 0xDC00);
-                }
-            }
-
-            if (firstOnly)
-            {
-                break;
-            }
-        }
-
-        projected = characters is null ? value : new(characters);
-        return true;
-    }
-
-    static CaseProjection GetProjection(string value)
-    {
-        if (value.Length == 0)
-        {
-            return CaseProjection.Exact;
-        }
-
-        if (IsAscii(value))
-        {
-            var hasLower = false;
-            var hasUpper = false;
-            var hasUpperAfterFirst = false;
-            for (var index = 0; index < value.Length; index++)
-            {
-                var character = value[index];
-                hasLower |= character is >= 'a' and <= 'z';
-                if (character is >= 'A' and <= 'Z')
-                {
-                    hasUpper = true;
-                    hasUpperAfterFirst |= index > 0;
-                }
-            }
-
-            if (!hasUpper)
-            {
-                return CaseProjection.Lower;
-            }
-
-            if (!hasLower)
-            {
-                return CaseProjection.Upper;
-            }
-
-            return value[0] is >= 'A' and <= 'Z' && !hasUpperAfterFirst
-                ? CaseProjection.Title
-                : CaseProjection.Mixed;
-        }
-
-        var hasUnicodeLower = false;
-        var hasUnicodeUpper = false;
-        var hasUnicodeUpperAfterFirst = false;
-        var firstScalarIsUpper = false;
-        for (var index = 0; index < value.Length;)
-        {
-            var isFirstScalar = index == 0;
-            var scalar = ReadScalar(value, ref index, value.Length);
-            var unicodeCase = InflectionUnicodeData.GetCase(scalar);
-            if (unicodeCase == InflectionUnicodeCase.Title)
-            {
-                return CaseProjection.Mixed;
-            }
-
-            if (unicodeCase == InflectionUnicodeCase.Lower)
-            {
-                hasUnicodeLower = true;
-            }
-            else if (unicodeCase == InflectionUnicodeCase.Upper)
-            {
-                hasUnicodeUpper = true;
-                firstScalarIsUpper |= isFirstScalar;
-                hasUnicodeUpperAfterFirst |= !isFirstScalar;
-            }
-        }
-
-        if (!hasUnicodeUpper)
-        {
-            return CaseProjection.Lower;
-        }
-
-        if (!hasUnicodeLower)
-        {
-            return CaseProjection.Upper;
-        }
-
-        return firstScalarIsUpper && !hasUnicodeUpperAfterFirst
-            ? CaseProjection.Title
-            : CaseProjection.Mixed;
-    }
-
-    bool IsEligibleToken(
-        string value,
-        out InflectionUnicodeScripts detectedScripts)
-    {
-        var comparison = casing == InflectionCasing.LowerTitleUpper
-            ? StringComparison.OrdinalIgnoreCase
-            : StringComparison.Ordinal;
-        if (Contains(skipSimpleWords, value, comparison))
-        {
-            detectedScripts = InflectionUnicodeScripts.None;
-            return false;
-        }
-
-        return TryGetDetectedScripts(
-            value,
-            allowNonLetters: false,
-            out detectedScripts);
-    }
-
-    bool IsScriptEligible(string value, bool allowNonLetters)
-        => TryGetDetectedScripts(value, allowNonLetters, out _);
-
-    bool TryGetDetectedScripts(
-        string value,
-        bool allowNonLetters,
-        out InflectionUnicodeScripts detectedScripts)
-    {
-        var allowedScripts = InflectionUnicodeScripts.None;
-        foreach (var script in scripts)
-        {
-            if (!InflectionUnicodeData.TryGetScript(script, out var scriptValue))
-            {
-                detectedScripts = InflectionUnicodeScripts.None;
-                return false;
-            }
-
-            allowedScripts |= scriptValue;
-        }
-
-        detectedScripts = allowedScripts;
-        var hasLetter = false;
-        for (var index = 0; index < value.Length; index++)
-        {
-            var scalar = char.ConvertToUtf32(value, index);
-            var category = CharUnicodeInfo.GetUnicodeCategory(value, index);
-            if (char.IsHighSurrogate(value[index]))
-            {
-                index++;
-            }
-
-            var pinnedLetter = false;
-            var pinnedMark = false;
-            if (category == UnicodeCategory.OtherNotAssigned)
-            {
-                _ = InflectionUnicodeData.TryGetPinnedLetterOrMark(
-                    scalar,
-                    out pinnedLetter,
-                    out pinnedMark);
-            }
-
-            var isLetter = pinnedLetter ||
-                category is UnicodeCategory.UppercaseLetter or
-                UnicodeCategory.LowercaseLetter or
-                UnicodeCategory.TitlecaseLetter or
-                UnicodeCategory.ModifierLetter or
-                UnicodeCategory.OtherLetter;
-            var isMark = pinnedMark ||
-                category is UnicodeCategory.NonSpacingMark or
-                UnicodeCategory.SpacingCombiningMark or
-                UnicodeCategory.EnclosingMark;
-            if (!isLetter && !isMark)
-            {
-                if (allowNonLetters)
-                {
-                    continue;
-                }
-
-                detectedScripts = InflectionUnicodeScripts.None;
-                return false;
-            }
-
-            var scalarScripts = InflectionUnicodeData.GetScripts(scalar);
-            if ((scalarScripts & allowedScripts) == InflectionUnicodeScripts.None)
-            {
-                detectedScripts = InflectionUnicodeScripts.None;
-                return false;
-            }
-
-            detectedScripts &= scalarScripts;
-            if (detectedScripts == InflectionUnicodeScripts.None)
-            {
-                return false;
-            }
-
-            hasLetter |= isLetter;
-        }
-
-        return hasLetter;
-    }
-
     bool TryCreateReverseCandidate(
         InflectionRule rule,
         string value,
@@ -1485,74 +781,6 @@ sealed class InflectionBundle
 
         stem = value.Substring(0, stemLength);
         return true;
-    }
-
-    static bool IsWellFormedUtf16(string value)
-    {
-        for (var index = 0; index < value.Length; index++)
-        {
-            if (char.IsHighSurrogate(value[index]))
-            {
-                if (++index >= value.Length || !char.IsLowSurrogate(value[index]))
-                {
-                    return false;
-                }
-            }
-            else if (char.IsLowSurrogate(value[index]))
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    static bool IsDefinitelyNormalized(string value)
-    {
-        for (var index = 0; index < value.Length; index++)
-        {
-            var scalar = char.ConvertToUtf32(value, index);
-            var category = CharUnicodeInfo.GetUnicodeCategory(value, index);
-            if (category is UnicodeCategory.NonSpacingMark or
-                UnicodeCategory.SpacingCombiningMark or
-                UnicodeCategory.EnclosingMark ||
-                HasUncertainNfcQuickCheck(scalar))
-            {
-                return false;
-            }
-
-            if (char.IsHighSurrogate(value[index]))
-            {
-                index++;
-            }
-        }
-
-        return true;
-    }
-
-    static bool HasUncertainNfcQuickCheck(int scalar)
-    {
-        var low = 0;
-        var high = UncertainNfcQuickCheckRanges.Length - 1;
-        while (low <= high)
-        {
-            var middle = low + ((high - low) / 2);
-            var range = UncertainNfcQuickCheckRanges[middle];
-            if (scalar < range.First)
-            {
-                high = middle - 1;
-            }
-            else if (scalar > range.Last)
-            {
-                low = middle + 1;
-            }
-            else
-            {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     readonly struct ReverseCandidate(
@@ -1819,133 +1047,6 @@ sealed class InflectionBundle
         return true;
     }
 
-    // <generated-unicode-nfc-quick-check>
-    // Unicode 16.0.0 NFC_QC=No/Maybe ranges from DerivedNormalizationProps.txt
-    // (SHA-256 4d4c03892dea9146d674b686e495df2d55a28d071ac474041d73518f887abddc).
-    static readonly InflectionUnicodeRange[] UncertainNfcQuickCheckRanges =
-    [
-        new(0x0300, 0x0304),
-        new(0x0306, 0x030C),
-        new(0x030F, 0x030F),
-        new(0x0311, 0x0311),
-        new(0x0313, 0x0314),
-        new(0x031B, 0x031B),
-        new(0x0323, 0x0328),
-        new(0x032D, 0x032E),
-        new(0x0330, 0x0331),
-        new(0x0338, 0x0338),
-        new(0x0340, 0x0345),
-        new(0x0374, 0x0374),
-        new(0x037E, 0x037E),
-        new(0x0387, 0x0387),
-        new(0x0653, 0x0655),
-        new(0x093C, 0x093C),
-        new(0x0958, 0x095F),
-        new(0x09BE, 0x09BE),
-        new(0x09D7, 0x09D7),
-        new(0x09DC, 0x09DD),
-        new(0x09DF, 0x09DF),
-        new(0x0A33, 0x0A33),
-        new(0x0A36, 0x0A36),
-        new(0x0A59, 0x0A5B),
-        new(0x0A5E, 0x0A5E),
-        new(0x0B3E, 0x0B3E),
-        new(0x0B56, 0x0B57),
-        new(0x0B5C, 0x0B5D),
-        new(0x0BBE, 0x0BBE),
-        new(0x0BD7, 0x0BD7),
-        new(0x0C56, 0x0C56),
-        new(0x0CC2, 0x0CC2),
-        new(0x0CD5, 0x0CD6),
-        new(0x0D3E, 0x0D3E),
-        new(0x0D57, 0x0D57),
-        new(0x0DCA, 0x0DCA),
-        new(0x0DCF, 0x0DCF),
-        new(0x0DDF, 0x0DDF),
-        new(0x0F43, 0x0F43),
-        new(0x0F4D, 0x0F4D),
-        new(0x0F52, 0x0F52),
-        new(0x0F57, 0x0F57),
-        new(0x0F5C, 0x0F5C),
-        new(0x0F69, 0x0F69),
-        new(0x0F73, 0x0F73),
-        new(0x0F75, 0x0F76),
-        new(0x0F78, 0x0F78),
-        new(0x0F81, 0x0F81),
-        new(0x0F93, 0x0F93),
-        new(0x0F9D, 0x0F9D),
-        new(0x0FA2, 0x0FA2),
-        new(0x0FA7, 0x0FA7),
-        new(0x0FAC, 0x0FAC),
-        new(0x0FB9, 0x0FB9),
-        new(0x102E, 0x102E),
-        new(0x1161, 0x1175),
-        new(0x11A8, 0x11C2),
-        new(0x1B35, 0x1B35),
-        new(0x1F71, 0x1F71),
-        new(0x1F73, 0x1F73),
-        new(0x1F75, 0x1F75),
-        new(0x1F77, 0x1F77),
-        new(0x1F79, 0x1F79),
-        new(0x1F7B, 0x1F7B),
-        new(0x1F7D, 0x1F7D),
-        new(0x1FBB, 0x1FBB),
-        new(0x1FBE, 0x1FBE),
-        new(0x1FC9, 0x1FC9),
-        new(0x1FCB, 0x1FCB),
-        new(0x1FD3, 0x1FD3),
-        new(0x1FDB, 0x1FDB),
-        new(0x1FE3, 0x1FE3),
-        new(0x1FEB, 0x1FEB),
-        new(0x1FEE, 0x1FEF),
-        new(0x1FF9, 0x1FF9),
-        new(0x1FFB, 0x1FFB),
-        new(0x1FFD, 0x1FFD),
-        new(0x2000, 0x2001),
-        new(0x2126, 0x2126),
-        new(0x212A, 0x212B),
-        new(0x2329, 0x232A),
-        new(0x2ADC, 0x2ADC),
-        new(0x3099, 0x309A),
-        new(0xF900, 0xFA0D),
-        new(0xFA10, 0xFA10),
-        new(0xFA12, 0xFA12),
-        new(0xFA15, 0xFA1E),
-        new(0xFA20, 0xFA20),
-        new(0xFA22, 0xFA22),
-        new(0xFA25, 0xFA26),
-        new(0xFA2A, 0xFA6D),
-        new(0xFA70, 0xFAD9),
-        new(0xFB1D, 0xFB1D),
-        new(0xFB1F, 0xFB1F),
-        new(0xFB2A, 0xFB36),
-        new(0xFB38, 0xFB3C),
-        new(0xFB3E, 0xFB3E),
-        new(0xFB40, 0xFB41),
-        new(0xFB43, 0xFB44),
-        new(0xFB46, 0xFB4E),
-        new(0x110BA, 0x110BA),
-        new(0x11127, 0x11127),
-        new(0x1133E, 0x1133E),
-        new(0x11357, 0x11357),
-        new(0x113B8, 0x113B8),
-        new(0x113BB, 0x113BB),
-        new(0x113C2, 0x113C2),
-        new(0x113C5, 0x113C5),
-        new(0x113C7, 0x113C9),
-        new(0x114B0, 0x114B0),
-        new(0x114BA, 0x114BA),
-        new(0x114BD, 0x114BD),
-        new(0x115AF, 0x115AF),
-        new(0x11930, 0x11930),
-        new(0x1611E, 0x16129),
-        new(0x16D67, 0x16D68),
-        new(0x1D15E, 0x1D164),
-        new(0x1D1BB, 0x1D1C0),
-        new(0x2F800, 0x2FA1D),
-    ];
-    // </generated-unicode-nfc-quick-check>
-
     static int FoldSimpleCase(int scalar) =>
         InflectionUnicodeData.FoldSimpleCase(scalar);
 
@@ -1969,19 +1070,4 @@ sealed class InflectionBundle
             CardinalPluralCategory.Other => InflectionExactRole.Other,
             _ => InflectionExactRole.None
         };
-
-    readonly struct InflectionUnicodeRange(int first, int last)
-    {
-        public int First { get; } = first;
-        public int Last { get; } = last;
-    }
-
-    enum CaseProjection
-    {
-        Exact,
-        Lower,
-        Title,
-        Upper,
-        Mixed
-    }
 }
