+++ src/Humanizer/Inflections/InflectionUnicodeData.cs
@@ -128,6 +128,47 @@ Ir6bBYDOAvJAwCC6A+ABglqQLeB0wDraCZAYugiAEJRN0Ca+QbDbK98D
         SimpleCaseFoldRangeCount);
     // </generated-unicode-case-fold>
 
+    // <generated-unicode-lowercase>
+    // Unicode 16.0.0 UnicodeData.txt simple lowercase mappings, compressed
+    // into 674 source ranges from 1,460 mappings. Source SHA-256:
+    // ff58e5823bd095166564a006e47d111130813dcf8bf234ef79fa51a870edb48f.
+    const int SimpleLowercaseRangeCount = 674;
+    const string SimpleLowercaseData =
+        "QRlAfxZAGAZAKAACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgAC" +
+        "AgACAgACAgCNAwIAAgIAAgIAAgMAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgMAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA8QEBAAICAAICAAIEAKQDAQACAgACAgCcAwEAAgIBmgMCAAID" +
+        "AJ4BAQCUAwEAlgMBAAICAJoDAQCeAwIApgMBAKIDAQACBACmAwEAqgMCAKwDAQACAgACAgACAgC0AwEAAgIAtAMDAAICALQDAQAC" +
+        "AgGyAwIAAgIAAgIAtgMBAAIEAAIIAAQBAAICAAQBAAICAAQBAAICAAICAAICAAICAAICAAICAAICAAICAAIDAAICAAICAAICAAIC" +
+        "AAICAAICAAICAAICAAIDAAQBAAICAAICAMEBAQBvAQACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgAC" +
+        "AgACAgACAgACAgACAgACAgCDAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAggA1qgBAQACAgDFAgEA0KgBAwACAgCFAwEAigEB" +
+        "AI4BAQACAgACAgACAgACAgACogIAAgIAAgQAAgkA6AEHAEwCAkoEAIABAgF+AxBAEghALAAQCQACAgACAgACAgACAgACAgACAgAC" +
+        "AgACAgACAgACAgACAgACBgB3AwACAgANAQACAwKDAgMPoAEQH0BQAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIC" +
+        "AAICAAICAAICAAICAAIKAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIC" +
+        "AAICAAICAAICAAICAAICAAICAAICAB4BAAICAAICAAICAAICAAICAAICAAIDAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIC" +
+        "AAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIC" +
+        "AAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIDJWDvFiXAcScAwHEGAMBx0wVPoN8EUAUQmREAAgcq/y4tAv8u" +
+        "wwIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgoA/XYCAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIC" +
+        "AAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAAIK" +
+        "Bw8QBQ8QBw8QBw8QBQ8RAA8CAA8CAA8CAA8JBw8gBw8QBw8QBw8QAQ8CAZMBAgARDAOrAQQAEQwBDwIBxwEOAQ8CAd8BAgANDAH/" +
+        "AQIB+wECABGqAgC5dQQA/YIBAQCLgQEHADguDyAjAAKzBhk0yg4vYGAAAgIA7acBAQDLOwEAzacBAwACAgACAgACAgC3qAEBAPmn" +
+        "AQEAvagBAQC7qAECAAIDAAIJAf2oAQIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgkAAgIAAgUAAs7yAQACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgAC" +
+        "AgACAgACAgACAgACAgACAgACAgACAgACAgACFAACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACAgACiAEAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgQAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAAgsAAgIAAgIAh6gEAQACAgACAgACAgACAgACBQACAgDPlAUDAAICAAIE" +
+        "AAICAAICAAICAAICAAICAAICAAICAAICAAICAAICAIeVBQEAnZUFAQCVlQUBAIGVBQEAh5UFAgCjlAUBANOUBQEAqZQFAQDADgEA" +
+        "AgIAAgIAAgIAAgIAAgIAAgIAAgIAAgIAXwEAhZUFAQDvqAQBAAICAAICAM2VBQEAAgQAAgYAAgIAAgIAAgIAgZkFGQACrK4BGUDf" +
+        "CSdQsAEjUMABCk4MDk4QBk4IAU7sDTKAAdABFUDQFh9AoKsBH0DA9QEhRA==";
+
+    static readonly int[] SimpleLowercaseRanges = DecodeMappingRanges(
+        SimpleLowercaseData,
+        SimpleLowercaseRangeCount);
+    // </generated-unicode-lowercase>
+
     // <generated-unicode-uppercase>
     // Unicode 16.0.0 UnicodeData.txt simple uppercase mappings, compressed
     // into 690 source ranges from 1,477 mappings. Source SHA-256:
@@ -281,6 +322,9 @@ AQABtQ4JAQsTARoFAdsTIQIiIQE=
 
     internal static int FoldSimpleCase(int scalar) =>
         MapSimple(scalar, SimpleCaseFoldRanges, SimpleCaseFoldRangeCount);
+
+    internal static int ToLowerSimple(int scalar) =>
+        MapSimple(scalar, SimpleLowercaseRanges, SimpleLowercaseRangeCount);
 
     internal static int ToUpperSimple(int scalar) =>
         MapSimple(scalar, SimpleUppercaseRanges, SimpleUppercaseRangeCount);
@@ -1090,5 +1134,164 @@ AQABtQ4JAQsTARoFAdsTIQIiIQE=
         while ((current & 0x80) != 0);
 
         return value;
+    }
+
+    internal static bool HasUncertainNfcQuickCheck(int scalar)
+    {
+        var low = 0;
+        var high = UncertainNfcQuickCheckRanges.Length - 1;
+        while (low <= high)
+        {
+            var middle = low + ((high - low) / 2);
+            var range = UncertainNfcQuickCheckRanges[middle];
+            if (scalar < range.First)
+            {
+                high = middle - 1;
+            }
+            else if (scalar > range.Last)
+            {
+                low = middle + 1;
+            }
+            else
+            {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+
+    // <generated-unicode-nfc-quick-check>
+    // Unicode 16.0.0 NFC_QC=No/Maybe ranges from DerivedNormalizationProps.txt
+    // (SHA-256 4d4c03892dea9146d674b686e495df2d55a28d071ac474041d73518f887abddc).
+    static readonly InflectionUnicodeRange[] UncertainNfcQuickCheckRanges =
+    [
+        new(0x0300, 0x0304),
+        new(0x0306, 0x030C),
+        new(0x030F, 0x030F),
+        new(0x0311, 0x0311),
+        new(0x0313, 0x0314),
+        new(0x031B, 0x031B),
+        new(0x0323, 0x0328),
+        new(0x032D, 0x032E),
+        new(0x0330, 0x0331),
+        new(0x0338, 0x0338),
+        new(0x0340, 0x0345),
+        new(0x0374, 0x0374),
+        new(0x037E, 0x037E),
+        new(0x0387, 0x0387),
+        new(0x0653, 0x0655),
+        new(0x093C, 0x093C),
+        new(0x0958, 0x095F),
+        new(0x09BE, 0x09BE),
+        new(0x09D7, 0x09D7),
+        new(0x09DC, 0x09DD),
+        new(0x09DF, 0x09DF),
+        new(0x0A33, 0x0A33),
+        new(0x0A36, 0x0A36),
+        new(0x0A59, 0x0A5B),
+        new(0x0A5E, 0x0A5E),
+        new(0x0B3E, 0x0B3E),
+        new(0x0B56, 0x0B57),
+        new(0x0B5C, 0x0B5D),
+        new(0x0BBE, 0x0BBE),
+        new(0x0BD7, 0x0BD7),
+        new(0x0C56, 0x0C56),
+        new(0x0CC2, 0x0CC2),
+        new(0x0CD5, 0x0CD6),
+        new(0x0D3E, 0x0D3E),
+        new(0x0D57, 0x0D57),
+        new(0x0DCA, 0x0DCA),
+        new(0x0DCF, 0x0DCF),
+        new(0x0DDF, 0x0DDF),
+        new(0x0F43, 0x0F43),
+        new(0x0F4D, 0x0F4D),
+        new(0x0F52, 0x0F52),
+        new(0x0F57, 0x0F57),
+        new(0x0F5C, 0x0F5C),
+        new(0x0F69, 0x0F69),
+        new(0x0F73, 0x0F73),
+        new(0x0F75, 0x0F76),
+        new(0x0F78, 0x0F78),
+        new(0x0F81, 0x0F81),
+        new(0x0F93, 0x0F93),
+        new(0x0F9D, 0x0F9D),
+        new(0x0FA2, 0x0FA2),
+        new(0x0FA7, 0x0FA7),
+        new(0x0FAC, 0x0FAC),
+        new(0x0FB9, 0x0FB9),
+        new(0x102E, 0x102E),
+        new(0x1161, 0x1175),
+        new(0x11A8, 0x11C2),
+        new(0x1B35, 0x1B35),
+        new(0x1F71, 0x1F71),
+        new(0x1F73, 0x1F73),
+        new(0x1F75, 0x1F75),
+        new(0x1F77, 0x1F77),
+        new(0x1F79, 0x1F79),
+        new(0x1F7B, 0x1F7B),
+        new(0x1F7D, 0x1F7D),
+        new(0x1FBB, 0x1FBB),
+        new(0x1FBE, 0x1FBE),
+        new(0x1FC9, 0x1FC9),
+        new(0x1FCB, 0x1FCB),
+        new(0x1FD3, 0x1FD3),
+        new(0x1FDB, 0x1FDB),
+        new(0x1FE3, 0x1FE3),
+        new(0x1FEB, 0x1FEB),
+        new(0x1FEE, 0x1FEF),
+        new(0x1FF9, 0x1FF9),
+        new(0x1FFB, 0x1FFB),
+        new(0x1FFD, 0x1FFD),
+        new(0x2000, 0x2001),
+        new(0x2126, 0x2126),
+        new(0x212A, 0x212B),
+        new(0x2329, 0x232A),
+        new(0x2ADC, 0x2ADC),
+        new(0x3099, 0x309A),
+        new(0xF900, 0xFA0D),
+        new(0xFA10, 0xFA10),
+        new(0xFA12, 0xFA12),
+        new(0xFA15, 0xFA1E),
+        new(0xFA20, 0xFA20),
+        new(0xFA22, 0xFA22),
+        new(0xFA25, 0xFA26),
+        new(0xFA2A, 0xFA6D),
+        new(0xFA70, 0xFAD9),
+        new(0xFB1D, 0xFB1D),
+        new(0xFB1F, 0xFB1F),
+        new(0xFB2A, 0xFB36),
+        new(0xFB38, 0xFB3C),
+        new(0xFB3E, 0xFB3E),
+        new(0xFB40, 0xFB41),
+        new(0xFB43, 0xFB44),
+        new(0xFB46, 0xFB4E),
+        new(0x110BA, 0x110BA),
+        new(0x11127, 0x11127),
+        new(0x1133E, 0x1133E),
+        new(0x11357, 0x11357),
+        new(0x113B8, 0x113B8),
+        new(0x113BB, 0x113BB),
+        new(0x113C2, 0x113C2),
+        new(0x113C5, 0x113C5),
+        new(0x113C7, 0x113C9),
+        new(0x114B0, 0x114B0),
+        new(0x114BA, 0x114BA),
+        new(0x114BD, 0x114BD),
+        new(0x115AF, 0x115AF),
+        new(0x11930, 0x11930),
+        new(0x1611E, 0x16129),
+        new(0x16D67, 0x16D68),
+        new(0x1D15E, 0x1D164),
+        new(0x1D1BB, 0x1D1C0),
+        new(0x2F800, 0x2FA1D),
+    ];
+    // </generated-unicode-nfc-quick-check>
+
+    readonly struct InflectionUnicodeRange(int first, int last)
+    {
+        public int First { get; } = first;
+        public int Last { get; } = last;
     }
 }
